refactor: investigate scanner Fixes #61
